Move selection functions from processor to selection.py, and some other cleanups
